+++ backend/src/main/java/com/vuthevy1209/springmail/service/embedding/EmbeddingService.java
@@ -3,6 +3,6 @@ package com.vuthevy1209.springmail.service.embedding;
 import java.util.List;
 
 public interface EmbeddingService {
-    List<Double> embed(String text);
-    List<List<Double>> embedBatch(List<String> texts);
+    List<Float> embed(String text);
+    List<List<Float>> embedBatch(List<String> texts);
 }
